requirements.txt
streamlit>=1.25.0
openai>=0.27.9
python-dotenv>=1.0.0
spacy>=3.6.0
en-core-web-sm 
